Share binary test models
